数据调整

+++ src/Program.cs
@@ -10,6 +10,8 @@ namespace src
     {
         static void Main(string[] args)
         {
+            var IsAdjust = true;
+
             var path = @"F:\基于买方意向的货物撮合交易\data\";
             if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
             {
@@ -17,11 +19,17 @@ namespace src
             }
 
             System.Text.Encoding.RegisterProvider(System.Text.CodePagesEncodingProvider.Instance);
+
+            if (IsAdjust){
+                Adjust.Run(path);    
+                return;
+            }
+
+
             var sellers = Seller.ReadSellerFile(path + "seller.csv");
             var buyers = Buyer.ReadBuyerFile(path + "buyer.csv");
             if (System.IO.File.Exists(path + "result.csv")) System.IO.File.Delete(path + "result.csv");
             //按照品种进行分组
-
             Parallel.ForEach(sellers.Select(x => x.品种).Distinct(), breed =>
             {
                 //if (breed == "SR")  //仅对SR测试
@@ -38,41 +46,7 @@ namespace src
                     System.Console.WriteLine("卖家数：" + sellers_Breed.Count);
                     System.Console.WriteLine("买家数：" + buyers_Breed.Count);
                     List<Result> results = Assign(sellers_Breed, buyers_Breed);
-                    results.Sort((x, y) => { return (x.买方客户 + x.卖方客户).CompareTo(y.买方客户 + y.卖方客户); });
-
-                    int hope_score = results.Sum(x => x.hope_score);
-                    //按照卖方进行GroupBy，然后Distinct仓库号
-                    var t = results.GroupBy(x => x.买方客户).Select(y => new { 品种 = y.First().品种, 仓库数 = y.Select(z => z.仓库).Distinct().Count() });
-                    int Diary_score = t.Sum(x =>
-                    {
-                        int score = 100;
-                        if (x.品种 == Utility.strCF)
-                        {
-                            score -= (x.仓库数 - 1) * 20;
-                        }
-                        else
-                        {
-                            score -= (x.仓库数 - 1) * 25;
-                        }
-                        return score;
-                    });
-                    System.Console.WriteLine("==============================================================");
-                    System.Console.WriteLine("总体贸易分单数：" + results.Count);
-                    System.Console.WriteLine("==============================================================");
-                    int score = (int)(hope_score * 0.6 + Diary_score * 0.4);
-                    System.Console.WriteLine("获得意向分数：" + hope_score);
-                    int totalhopescore = buyers_Breed.Sum(x => x.TotalHopeScore);
-                    System.Console.WriteLine("最大意向分数：" + totalhopescore);
-                    System.Console.WriteLine("意向得分率：" + (hope_score * 100 / totalhopescore) + "%");
-                    System.Console.WriteLine("==============================================================");
-                    System.Console.WriteLine("记录分数：" + Diary_score);
-                    System.Console.WriteLine("最大记录分数：" + 100 * buyers_Breed.Count);
-                    System.Console.WriteLine("记录得分率：" + (Diary_score / buyers_Breed.Count) + "%");
-                    System.Console.WriteLine("==============================================================");
-                    System.Console.WriteLine("总体分数：" + score);
-                    int score_stardard = buyers_Breed.Count * 100;
-                    System.Console.WriteLine("得分率：" + (score * 100 / score_stardard) + "%");
-                    System.Console.WriteLine("==============================================================");
+                    Result.Score(results,buyers_Breed);
                     Result.AppendToCSV(path + "result.csv", results);
                 }
             });
